+++ cmake/SuperBuild.cmake
@@ -22,7 +22,7 @@
 # THE SOFTWARE.
 ################################################################################
 
-cmake_minimum_required(VERSION 3.5)
+cmake_minimum_required(VERSION 3.10)
 
 include(GNUInstallDirs)
 include(ExternalProject)