+++ pdi/cmake/FindGMock.cmake
@@ -189,7 +189,7 @@
 #
 #     add_test(AllTestsInFoo foo)
 
-cmake_minimum_required(VERSION 3.5)
+cmake_minimum_required(VERSION 3.10)
 list(INSERT CMAKE_MODULE_PATH 0 "${CMAKE_CURRENT_LIST_DIR}")
 
 function(__gmock_append_debugs _endvar _library)