+++ plugins/set_value/CHANGELOG.md
@@ -7,11 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
-* Support for changing logger level and pattern during execution.
-  [#363](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/363)
 
 ### Changed
-* Updated dependencies, now require cmake-3.10.
 
 ### Deprecated
 
@@ -20,6 +17,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 ### Security
+
+
+## [1.5.0] - 2022-03-30
+
+### Added
+* Support for changing logger level and pattern during execution.
+  [#363](https://gitlab.maisondelasimulation.fr/pdidev/pdi/-/issues/363)
+
+### Changed
+* Updated dependencies, now require cmake-3.10.
 
 
 ## [1.3.0] - 2021-07-30